relized adjoint method of backpropagation
